+++ app/js/utils/notify.js
@@ -1,0 +1,20 @@
+var notify = {
+    // is there but isn't used anymore
+    bottom: function (message, duration, anim, animduration) {
+        if (!duration) duration = 2500;
+        $('#notification').html(message).delay(0).queue(function () {
+            if (anim && animduration) interface.animate('#notification', anim, animduration);
+            $('#notification').dequeue();
+        }).delay(duration).queue(function () {
+            $('#notification').html('').dequeue();
+        });
+    },
+    
+    // top-right notification sliding
+    snack: function (message, duration) {
+        if (!duration) duration = 2500;
+        $('#notification-snack').html(message).show().addClass('slideNotification').delay(duration).queue(function () {
+            $('#notification-snack').html('').hide('fast').removeClass('slideNotification').dequeue();
+        });
+    }
+};
