+++ app/js/utils/os-lang.json
