+++ app/localization/en.json
@@ -1,5 +1,5 @@
 {
-    "currentLang": "English",
+	"currentLang": "English",
 	"Upload": "Upload",
 	"Log in": "Log in",
 	"Log out": "Log out",
