+++ app/localization/nl.json
@@ -21,7 +21,7 @@
 	"Size": "Grootte",
 	"bytes": "bytes",
 	"High definition": "High definition",
-	"Movie AKA": "Film aliasnaam",
+	"Movie AKA": "Alternatieve titel",
 	"Release name": "Releasenaam",
 	"FPS": "FPS",
 	"Total time": "Tijdsduur",
