+++ app/localization/ro.json
@@ -1,5 +1,5 @@
 {
-    "currentLang": "Română",
+	"currentLang": "Română",
 	"Upload": "Încarcă",
 	"Log in": "Înregistrare",
 	"Log out": "Deconectare",
